engine: remove dependency on rules, fixing circular import
